Added 'make portagepatch' support
This produces a patch I can send to Joe. This produces differences just
for the stage directory so this concentrates on the changes needed to
the portage folder only. This diffs the working tree's stage folder
against a previous formal git tag, so it first makes sure the stage
folder reflects the current working tree's version of the ebuild and the
patches are generated etc.