Moved client playlists as part of bug#416923
Client playlists are m3u files that are used to remember the state of
the player. LMS persists these when the playlist changes so that the
state of the player is restored in the server is restarted. Previously,
LMS would save the client playlists into the 'prefs' directory
(/etc/logitechmediaserver under this ebuild).

This commit changes that to be under /var/lib/ClientPlaylists since
these playlists aren't really preferences because they dynamically
change.