Bumped to alpha2 -- fix dependencies for amd64
The 7.7.3_alpha2 ebuild no longer is no longer keyworded to be available
to amd64. This is a temporary problem because the test tarball released
by Logitech does not include any binaries for amd64 for the 5.16 Perl
arch. I expect that will be fixed when a full Perl-5.16-compatible
version is released.